+++ CHANGELOG.md
@@ -70,6 +70,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/067]:** com.google.fonts/check/name_familyname_first_char
   - **[com.google.fonts/check/068]:** com.google.fonts/check/nameid/match_familyname_fullfont
   - **[com.google.fonts/check/069]:** com.google.fonts/check/glyf_unused_data
+  - **[com.google.fonts/check/070]:** com.google.fonts/check/currency_chars
   - **[com.google.fonts/check/071]:** com.google.fonts/check/family_naming_recommendations
   - **[com.google.fonts/check/073]:** com.google.fonts/check/maxadvancewidth
   - **[com.google.fonts/check/075]:** com.google.fonts/check/points_out_of_bounds